+++ SPaceInvader/src/main/java/be/uantwerpen/fti/ei/spaceinvaders/game/entity/entitycomponents/PositionComponent.java
@@ -31,19 +31,19 @@ public class PositionComponent {
     }
 
 
-    public int getX() {
+    public double getX() {
         return position.getX();
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.position.setX(x);
     }
 
-    public int getY() {
+    public double getY() {
         return this.position.getY();
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.position.setY(y);
     }
